The secret of freemasonry is that the Cainites are correct
Forgive them. I mean really fuck you guys but forgiveness is fuck you I love you so meh. You made everyone’s lives hell on earth. Good news is now we can have it on earth as it is in heaven. Then we get to go up to the stars. Life on planet earth was a test. The other celestial bodies don’t want life if it’s gonna be an issue. They’re happy spinning round for eternity to allow us to exist is by the grace of our Holy Earthly Mother. We must respect and honor Her and Him. Saying 22 will lead us to the stars.

Signed-off-by: John Andersen <johnandersenpdx@gmail.com>